+++ emborg/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.30.2"
-__released__ = "2022-02-15"
+__version__ = "1.30.3"
+__released__ = "2022-03-13"
 
 from .settings import Settings as Emborg